fix:2차 수정, 컴포넌트 나누기
