+++ src/components/myPage/MyPageContent.tsx
@@ -8,33 +8,13 @@ import MyProducts from "@/components/myPage/menu/myProducts";
 import Orders from "@/components/myPage/menu/order/Orders";
 import Profile from "@/components/myPage/menu/profile/Profile";
 import { useProductStore } from "@/store";
-import { IoIosArrowForward } from "react-icons/io";
-
 import MoreButton from "./MoreButton";
 import ProfileButton from "./ProfileButton";
 
-type MyPageContentProps = {
-  myPageMenuList: {
-    id: number;
-    icon: JSX.Element;
-    title: string;
-    value: string;
-  }[]
-}
-
-const MyPageContent = ({ myPageMenuList }: MyPageContentProps) => {
+const MyPageContent = () => {
   const { menuNumber } = useProductStore();
-
   return (
-    <div className="w-[900px]">
-      <div className={` -mt-8  ${menuNumber === 0 ? 'hidden' : 'block'}`}>
-        <div className="flex items-center mb-2">
-          <p className="text-[#9f9f9f]">{`myPage`}&nbsp;&nbsp;&nbsp;</p>
-          <IoIosArrowForward className="text-[20px] text-[#9f9f9f]" />
-          <p>&nbsp;&nbsp;&nbsp;</p>
-          <p className="text-white">{menuNumber === 0 ? '' : menuNumber === 6 ? 'MyProducts' : myPageMenuList[menuNumber - 1].value}</p>
-        </div>
-      </div>
+    <>
       {menuNumber === 0 ?
         <div className="w-[900px]">
           <div className="flex items-center justify-between w-full border rounded-xl px-8 py-4 bg-white mb-5">
@@ -79,7 +59,7 @@ const MyPageContent = ({ myPageMenuList }: MyPageContentProps) => {
                   ? <Profile />
                   : <MyProducts />
       }
-    </div>
+    </>
   )
 }
